requirements.txt
apache-airflow==2.6.0
apache-airflow-providers-slack
google-cloud-bigquery
google-cloud-storage
google-cloud-functions
apache-airflow-providers-google
dbt-core==1.5.0
dbt-bigquery==1.5.0
pandas
python-dotenv
pydantic<2.0.0